Backend/farmchainx-flask/requirements.txt
flask==3.0.3
flask-cors==4.0.1
python-dotenv==1.0.1
requests==2.32.3
inference-sdk>=0.9
